requirements.txt
APScheduler >= 3.7.0
PyAutoGUI =>= 0.9.52
PySimpleGUI >= 4.32.1
aiofile =>= 3.5.0
aiohttp >= 3.6.3
bs4 >= 0.0.1
captcha == 0.3
coc.py >= 1.1.0
discord_ext_menus >= 1.0.0a30+g309c702
emojis >= 0.6.0
fonttools >= 4.18.2
frida >= 14.2.15
google_trans_new >= 1.1.9
nest_asyncio >= 1.5.1
pillow >= 8.0.1
psutil >= 5.8.0
pure_python_adb >= 0.3.0.dev0
pyperclip >= 1.8.1
python_dateutil >= 2.8.1
python_dotenv >= 0.15.0
pywin32 >= 300
requests >= 2.25.0
translate >= 3.5.0
wavelink >= 0.9.9
